refactor: move search_fields
Move `search_fields` from `server.py` to `fields.py`.
